Moved tag handler to common module
